+++ test/e2e/helm_install_upgrade_test.go
@@ -3,6 +3,8 @@ package e2e
 import (
 	"context"
 	"fmt"
+	"os"
+	"path/filepath"
 	"strings"
 	"testing"
 	"time"
@@ -33,13 +35,19 @@ import (
 )
 
 func TestHelmUpgrade(t *testing.T) {
-	const (
-		// Rel: https://github.com/Kong/charts/tree/main/charts/kong-operator
-		chart = "kong/kong-operator"
-
-		waitTime = 3 * time.Minute
+	t.Skip(
+		"Skipping Helm upgrade test until we get images that contain the changes +",
+		"from https://github.com/Kong/kong-operator/pull/2201. "+
+			"That PR introduced a change in validating webhook certificates paths and "+
+			"Helm chart can only use 1 path for the certificates so it won't work for "+
+			"images before and after that change at the same time.",
 	)
-	ctx := t.Context()
+	const waitTime = 3 * time.Minute
+
+	var (
+		ctx      = t.Context()
+		chartDir = filepath.Join(testutils.ProjectRootPath(), "charts/kong-operator")
+	)
 
 	// createEnvironment will queue up environment cleanup if necessary
 	// and dumping diagnostics if the test fails.
@@ -139,7 +147,7 @@ func TestHelmUpgrade(t *testing.T) {
 		},
 		{
 			name:             "upgrade from latest minor to current",
-			fromVersion:      "2.0.0-alpha.4", // renovate: datasource=docker packageName=kong/kong-operator depName=kong/kong-operator
+			fromVersion:      "2.0.0-alpha.5", // renovate: datasource=docker packageName=kong/kong-operator depName=kong/kong-operator
 			upgradeToCurrent: true,
 			// This is the effective semver of a next release.
 			// It's needed for the chart to properly render semver-conditional templates.
@@ -361,11 +369,17 @@ func TestHelmUpgrade(t *testing.T) {
 					"upgrade": {
 						"--devel",
 					},
+					"uninstall": {
+						"--wait",
+					},
 				},
 			}
 
-			require.NoError(t, helm.AddRepoE(t, opts, "kong", "https://charts.konghq.com"))
-			require.NoError(t, helm.InstallE(t, opts, chart, releaseName))
+			dir, err := os.Getwd()
+			require.NoError(t, err)
+			t.Logf("dir: %s", dir)
+			t.Logf("Using Helm chart dir : %s", chartDir)
+			require.NoError(t, helm.InstallE(t, opts, chartDir, releaseName))
 			out, err := helm.RunHelmCommandAndGetOutputE(t, opts, "list")
 			require.NoError(t, err)
 			t.Logf("Helm list output after install:\n  %s", out)
@@ -383,7 +397,12 @@ func TestHelmUpgrade(t *testing.T) {
 			// Deploy the objects that should be present before the upgrade.
 			cl := client.NewNamespacedClient(e.Clients.MgrClient, e.Namespace.Name)
 			for _, obj := range tc.objectsToDeploy {
-				require.NoError(t, cl.Create(ctx, obj))
+				// NOTE: Create objects with eventually since we're deploying
+				// admission webhook and that can take a moment to become ready.
+				require.EventuallyWithT(t, func(t *assert.CollectT) {
+					obj := obj.DeepCopyObject().(client.Object)
+					require.NoError(t, cl.Create(ctx, obj))
+				}, waitTime, 500*time.Millisecond)
 				t.Cleanup(func() {
 					// Ensure that every object is properly deleted (the finalizer must
 					// be executed, it requires some time) before the Helm chart is uninstalled.
@@ -407,7 +426,7 @@ func TestHelmUpgrade(t *testing.T) {
 			opts.SetValues["image.tag"] = tag
 			opts.SetValues["image.repository"] = targetRepository
 
-			require.NoError(t, helm.UpgradeE(t, opts, chart, releaseName))
+			require.NoError(t, helm.UpgradeE(t, opts, chartDir, releaseName))
 			out, err = helm.RunHelmCommandAndGetOutputE(t, opts, "list")
 			require.NoError(t, err)
 			t.Logf("Helm list output after upgrade:\n  %s", out)